rename one of format files to fix issues with some compilers
