+++ src/historywork/ApplyBucketsWork.cpp
@@ -9,9 +9,9 @@
 #include "bucket/BucketManager.h"
 #include "crypto/Hex.h"
 #include "history/HistoryArchive.h"
+#include "historywork/Progress.h"
 #include "ledger/LedgerManager.h"
 #include "main/Application.h"
-#include "util/Format.h"
 #include "util/make_unique.h"
 
 namespace stellar
