+++ src/historywork/BatchDownloadWork.cpp
@@ -5,9 +5,9 @@
 #include "historywork/BatchDownloadWork.h"
 #include "history/HistoryManager.h"
 #include "historywork/GetAndUnzipRemoteFileWork.h"
+#include "historywork/Progress.h"
 #include "lib/util/format.h"
 #include "main/Application.h"
-#include "util/Format.h"
 
 namespace stellar
 {
