+++ src/historywork/WriteSnapshotWork.cpp
@@ -5,9 +5,9 @@
 #include "historywork/WriteSnapshotWork.h"
 #include "database/Database.h"
 #include "history/StateSnapshot.h"
+#include "historywork/Progress.h"
 #include "ledger/LedgerHeaderFrame.h"
 #include "main/Application.h"
-#include "util/Format.h"
 #include "util/XDRStream.h"
 
 namespace stellar
